+++ modules/bookings/components/BookingsPage/MainPage/style.scss
@@ -1,0 +1,40 @@
+.MainPage {
+    align-self: center;
+    color: var(--mid-gray);
+    text-align: center;
+    font-family: brother-1816, sans-serif;
+    font-style: normal;
+    font-weight: lighter;
+  
+    &_Title {
+      font-size: 2.5em;
+      font-weight: lighter;
+      margin-bottom: 0.5em;
+    }
+  
+    &_Instruction {
+      font-size: 0.8em;
+      margin-bottom: 1.5em;
+    }
+  
+    &_Button {
+      margin: 0.6em;
+      width: 100%;
+      text-align: center;
+  
+      >div {
+        display: inline-block;
+        width: 25%;
+        border: 1px solid rgba(var(--dark-gray), 1);
+        padding: 0.5em;
+        width: 25%;
+  
+        >a {
+          font-size: 0.7em;
+          font-weight: bold;
+          color: rgba(var(--dark-gray), 1);
+          font-family: brother-1816, sans-serif;
+        }
+      }
+    }
+  }
